Update form labels, display balance from localStorage, and add latest transactions to homepage
The form labels in TransferForm have been changed from 'Sender' to 'Sender-Nummer' and 'Empfänger' to 'Empfänger-Nummer'. The balance is now retrieved from localStorage in AccountTransactionPage and displayed in its current format. On the Home page, an overview of the latest transactions has been added. This will show up to five recent transactions, stored in localStorage.